bazarw_front_end/requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Django==3.2.4
djangorestframework==3.12.4
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==2.15.6
mysqlclient==2.0.3
pyrsistent==0.17.3
pytz==2021.1
requests==2.25.1
rests==0.0.1
six==1.16.0
sqlparse==0.4.1
transbank-sdk==1.5.0
urllib3==1.26.5
Werkzeug==2.0.1
